Revert "Machrihanish (MAC) [DME ONLY] reinstated, Newcastle (NEW) [DME ONLY] reinstated"
This reverts commit 82d3434b02c299d229fc51942ee3c6dbc5c40adc.

This wasn't a bug, they were intentionally commented out to remove them
from view in the sector. To an uneducated controller they could be
mistaken for VORs. I didn't think there was any reason to have them and
nobody has asked until now.

If there's a good argument for adding them we can look at doing that in a
future version.