Adjust default vis centre to an appropriate middle-of-the-UK compromise
Upon testing with EuroScope 3.2, it appears that the default visibility
centre is used to determine the lateral and longitudinal scale factors the
radar display, and the values given in the sector file [INFO] (60/36.06
here) are ignored. This latitude (53.03... N) corresponds to the value
36.06 NM per degree of longitude, which is a compromise that gives equal
distortion at 50N and 56N - roughly the limits of the area of interest for
UK domestic area controlling.

The longitude value (002W) has been chosen as a round number to make clear
that it has no other significance. It's roughly halfway across UK airspace
at this latitude, but also is situated conveniently between Manchester and
the midlands airports. In situations where it is used as the default
visibility centre, this means that users with low vis ranges are more
likely to be able to see something interesting.

Visually, it may seem that the UK stands a little bit taller than before.

Incidentally, the final [INFO] value (10 in this file), after the magnetic
variation, was defined as some kind of overall 'scale factor' but as far
as I can tell is not used in EuroScope, and as far as I know not in
other current controller clients.