Adjust STAR ordering for best default runway selection
When a flight plan route ends with a STAR with no runway specified (e.g. ...DOMUT
WILLO4C), and there is more than one runway available on this STAR,
EuroScope selects the first STAR/runway combination specified in the
sector file, regardless of the active runway.

While this clearly isn't ideal, to minimise the inconvenience it is better
to ensure that the most likely arrival runway is being used by default.
Especially for somewhere like Gatwick, with an unlikely spare runway, 26L
is a much better default than 26R.

This commit re-orders the STARs for best defaults, for those airports
where the current order was unsatisfactory.