Add stand outlines
Outlines are added to stands 1-18, 101 and 102. These are included in the main 'EGMC Southend' GEO category.
The centre-lines of these stands are moved to a separated 'EGMC Southend Extra' GEO category.
Stand numbers are re-positioned to the middle of the stand, rather than the nose-in end.
The South apron lines are moved to their current position and changed to 'smrWhite'.
A "South Apron" label is added.